Move the settings endpoint.
Move GET system/setting/large_image to GET large_image/settings to keep the namespace cleaner.
